+++ mysite/api/serializer.py
@@ -1,0 +1,10 @@
+from django.shortcuts import render
+from django.http import JsonResponse
+from django.contrib.auth import get_user_model
+from rest_framework import serializers
+
+
+class UsernameSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = get_user_model()
+        fields = ["id", 'username', ]