+++ mysite/calorietracker/base_models.py
@@ -1,4 +1,7 @@
 from django.db import models
+from measurement import measures
+from django.utils.deconstruct import deconstructible
+
 # from safedelete.models import SafeDeleteModel
 # from django.contrib.auth import get_user_model
 
@@ -11,3 +14,13 @@ class DateTimeFields(models.Model):
 
     class Meta:
         abstract = True
+
+
+@deconstructible
+class Weight(measures.Weight):
+    pass
+
+
+@deconstructible
+class Distance(measures.Distance):
+    pass