+++ mysite/calorietracker/forms.py
@@ -12,6 +12,18 @@ from cloudinary.forms import CloudinaryFileField
 from .models import Log, Setting, Image
 
 
+class FriendShipRequestForm(forms.Form):
+    """docstring for AcceptFriendForm."""
+
+    to_user = forms.ModelChoiceField(get_user_model().objects.all())
+
+
+class FriendForm(forms.Form):
+    """docstring for AcceptFriendForm."""
+
+    from_user = forms.ModelChoiceField(get_user_model().objects.all())
+
+
 class MeasurementWidget(forms.MultiWidget):
     def __init__(
         self,