+++ mysite/calorietracker/models.py
@@ -1,13 +1,14 @@
 import datetime
 
+from cloudinary.models import CloudinaryField
 from django.contrib.auth import get_user_model
 from django.db import models
 from django_cryptography.fields import encrypt
 from django_measurement.models import MeasurementField
-from measurement.measures import Distance, Weight
+from pinax.referrals.models import Referral
 from safedelete.models import SafeDeleteModel
 
-from .base_models import DateTimeFields
+from .base_models import DateTimeFields, Distance, Weight
 from pinax.referrals.models import Referral
 from cloudinary.models import CloudinaryField
 import uuid
@@ -20,11 +21,22 @@ class AnalyticsShareToken(DateTimeFields, SafeDeleteModel):
     uuid = models.UUIDField(default=uuid.uuid4)
 
 
+def get_default_goal_date():
+    return datetime.datetime.today() + datetime.timedelta(days=30)
+
+
+def get_default_goal_mfp_autosync_startdate():
+    return datetime.datetime.today() - datetime.timedelta(days=90)
+
+
 class Wallet(DateTimeFields, SafeDeleteModel):
     """docstring for Feedback."""
 
     user = models.OneToOneField(get_user_model(), on_delete=models.CASCADE)
     coins = models.IntegerField(default=0)
+
+    def __str__(self):
+        return f"Wallet from {self.user}: {self.coins}"
 
 
 class MFPCredentials(DateTimeFields, SafeDeleteModel):
@@ -33,6 +45,15 @@ class MFPCredentials(DateTimeFields, SafeDeleteModel):
     user = models.OneToOneField(get_user_model(), on_delete=models.CASCADE)
     username = models.CharField(max_length=255)
     password = encrypt(models.CharField(max_length=255))
+    mfp_autosync = models.BooleanField(default=False)
+    mfp_autosync_startdate = models.DateTimeField(
+        default=get_default_goal_mfp_autosync_startdate,
+        blank=True,
+    )
+    last_mfp_log_date_synced = models.DateTimeField(null=True, blank=True)
+
+    def __str__(self):
+        return f"MFPCredentials from {self.user}"
 
 
 class Feedback(DateTimeFields, SafeDeleteModel):
@@ -41,6 +62,9 @@ class Feedback(DateTimeFields, SafeDeleteModel):
     user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
     comment = models.TextField()
     contact_email = models.EmailField()
+
+    def __str__(self):
+        return f"Feedback from {self.user}: {self.comment[0:10]}"
 
 
 class Subscription(DateTimeFields, SafeDeleteModel):
@@ -131,7 +155,7 @@ class Setting(DateTimeFields, SafeDeleteModel):
         max_length=1, choices=sex_choices, blank=True, null=True, default="M"
     )
     height = MeasurementField(
-        measurement=Distance, null=True, blank=True, default=1.75
+        measurement=Distance, null=True, blank=True, default=Distance(m=1.75)
     )  # height in metres
 
     activity_choices = [
@@ -165,10 +189,10 @@ class Setting(DateTimeFields, SafeDeleteModel):
     )
 
     goal_weight = MeasurementField(
-        measurement=Weight, null=True, blank=True, default=80
+        measurement=Weight, null=True, blank=True, default=Weight(kg=80)
     )  # default weight is in kg
     goal_date = models.DateTimeField(
-        blank=True, null=True, default=datetime.datetime.now
+        blank=True, null=True, default=get_default_goal_date
     )
     unit_choices = [
         ("I", "Imperial (pounds, feet, inches etc.)"),
@@ -182,6 +206,9 @@ class Setting(DateTimeFields, SafeDeleteModel):
         help_text="Display metric or imperial units on analytics page",
         default="M",
     )
+
+    def __str__(self):
+        return f"Settings from {self.user}"
 
 
 class Image(DateTimeFields, SafeDeleteModel):
@@ -242,3 +269,6 @@ class Log(DateTimeFields, SafeDeleteModel):
     front_progress_pic = CloudinaryField("image", null=True, blank=True)
     side_progress_pic = CloudinaryField("image", null=True, blank=True)
     back_progress_pic = CloudinaryField("image", null=True, blank=True)
+
+    def __str__(self):
+        return f"Log from {self.user}"